Move option components in dedicated folder
