+++ packages/app/src/components/VisualizationOptions/Options/AggregationType.js
