+++ packages/app/src/components/VisualizationOptions/Options/CumulativeValues.js
