+++ packages/app/src/components/VisualizationOptions/Options/HideEmptyRowItems.js
