+++ packages/app/src/components/VisualizationOptions/Options/HideLegend.js
