+++ packages/app/src/components/VisualizationOptions/Options/HideSubtitle.js
