+++ packages/app/src/components/VisualizationOptions/Options/HideTitle.js
