+++ packages/app/src/components/VisualizationOptions/Options/NoSpaceBetweenColumns.js
