+++ packages/app/src/components/VisualizationOptions/Options/PercentStackedValues.js
