+++ packages/app/src/components/VisualizationOptions/Options/RangeAxisDecimals.js
