+++ packages/app/src/components/VisualizationOptions/Options/RangeAxisMaxValue.js
