+++ packages/app/src/components/VisualizationOptions/Options/RegressionType.js
