+++ packages/app/src/components/VisualizationOptions/Options/TargetLine.js
@@ -5,12 +5,12 @@ import FormGroup from '@material-ui/core/FormGroup';
 import Switch from '@material-ui/core/Switch';
 
 import i18n from '@dhis2/d2-i18n';
-import { sGetUiOptions } from '../../reducers/ui';
-import { acSetUiOptions } from '../../actions/ui';
-import BaseLineValue from './BaseLineValue';
-import BaseLineLabel from './BaseLineLabel';
+import { sGetUiOptions } from '../../../reducers/ui';
+import { acSetUiOptions } from '../../../actions/ui';
+import TargetLineValue from './TargetLineValue';
+import TargetLineLabel from './TargetLineLabel';
 
-const BaseLine = ({ enabled, onChange }) => (
+const TargetLine = ({ enabled, onChange }) => (
     <FormGroup row={true}>
         <FormControlLabel
             control={
@@ -19,26 +19,26 @@ const BaseLine = ({ enabled, onChange }) => (
                     onChange={event => onChange(event.target.checked)}
                 />
             }
-            label={i18n.t('Base line')}
+            label={i18n.t('Target line')}
         />
         {enabled ? (
             <Fragment>
-                <BaseLineValue />
-                <BaseLineLabel />
+                <TargetLineValue />
+                <TargetLineLabel />
             </Fragment>
         ) : null}
     </FormGroup>
 );
 
 const mapStateToProps = state => ({
-    enabled: sGetUiOptions(state).baseLine,
+    enabled: sGetUiOptions(state).targetLine,
 });
 
 const mapDispatchToProps = dispatch => ({
-    onChange: enabled => dispatch(acSetUiOptions({ baseLine: enabled })),
+    onChange: enabled => dispatch(acSetUiOptions({ targetLine: enabled })),
 });
 
 export default connect(
     mapStateToProps,
     mapDispatchToProps
-)(BaseLine);
+)(TargetLine);
